applications: Modernize C-style casts

+++ src/applications/model/bulk-send-application.cc
@@ -172,8 +172,8 @@ BulkSendApplication::SendData(const Address& from, const Address& to)
             packet = Create<Packet>(toSend);
         }
 
-        int actual = m_socket->Send(packet);
-        if (static_cast<uint64_t>(actual) == toSend)
+        uint32_t actual = m_socket->Send(packet);
+        if (actual == toSend)
         {
             m_totBytes += actual;
             m_txTrace(packet);
@@ -194,11 +194,9 @@ BulkSendApplication::SendData(const Address& from, const Address& to)
             // a quantity less than the packet size.  Split the packet
             // into two, trace the sent packet, save the unsent packet
             NS_LOG_DEBUG("Packet size: " << packet->GetSize() << "; sent: " << actual
-                                         << "; fragment saved: "
-                                         << toSend - static_cast<uint64_t>(actual));
+                                         << "; fragment saved: " << toSend - actual);
             Ptr<Packet> sent = packet->CreateFragment(0, actual);
-            Ptr<Packet> unsent =
-                packet->CreateFragment(actual, toSend - static_cast<uint64_t>(actual));
+            Ptr<Packet> unsent = packet->CreateFragment(actual, toSend - actual);
             m_totBytes += actual;
             m_txTrace(sent);
             m_unsentPacket = unsent;
